+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   * `hypot()`: given *n* numbers, outputs the length of the *n*-dimensional
     vector that has components equal to each of the inputs.
 
-* Add the variables `pi` and `e` to the built-in "sass:math" module.
+* Add the variables `$pi` and `$e` to the built-in "sass:math" module.
 
 ## 1.24.0